git-svn-id: https://svn.neo4j.org/components/kernel/branches/ha@4371 0b971d98-bb2f-0410-8247-b05b2b5feb2a

+++ community/src/main/java/org/neo4j/kernel/impl/nioneo/store/NeoStore.java
@@ -21,6 +21,8 @@ package org.neo4j.kernel.impl.nioneo.store;
 
 import java.io.IOException;
 import java.nio.ByteBuffer;
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Map;
 import java.util.Random;
 
@@ -119,6 +121,11 @@ public class NeoStore extends AbstractStore
 
     public void flushAll()
     {
+        if ( relTypeStore == null || propStore == null || relStore == null || 
+                nodeStore == null )
+        {
+            return;
+        }
         relTypeStore.flushAll();
         propStore.flushAll();
         relStore.flushAll();
@@ -371,5 +378,15 @@ public class NeoStore extends AbstractStore
     public int getRelationshipGrabSize()
     {
         return REL_GRAB_SIZE;
+    }
+    
+    public List<WindowPoolStats> getAllWindowPoolStats()
+    {
+        List<WindowPoolStats> list = new ArrayList<WindowPoolStats>();
+        list.addAll( nodeStore.getAllWindowPoolStats() );
+        list.addAll( propStore.getAllWindowPoolStats() );
+        list.addAll( relStore.getAllWindowPoolStats() );
+        list.addAll( relTypeStore.getAllWindowPoolStats() );
+        return list;
     }
 }
